pyproject.toml
[project]
name = "agentic-ai-mod-7-demo-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.97.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
